Split ulmfit/thai2fit's huge __init__.py into smaller files
